feat(FE): add pause button in trasaction history graphic
