+++ webapp/src/gql/index.js
@@ -1,2 +1,3 @@
 export * from './producer.gql'
 export * from './setting.gql'
+export * from './transaction.gql'
